+++ src/cjs/Polygon.js
@@ -418,8 +418,11 @@ var Polygon = /** @class */ (function () {
         center.y /= this.vertices.length;
         return center;
     };
+    //--- BEGIN --- Implement interface `Intersectable`
     /**
      * Get all line intersections with this polygon.
+     *
+     * This method returns all intersections (as vertices) with this shape. The returned array of vertices is in no specific order.
      *
      * See demo `47-closest-vector-projection-on-polygon` for how it works.
      *
@@ -450,6 +453,15 @@ var Polygon = /** @class */ (function () {
             .locateLineIntersecion(line, this.vertices, this.isOpen, inVectorBoundsOnly)
             .map(function (intersectionTuple) { return intersectionTuple.intersectionPoint; });
     };
+    /**
+     * Get all line intersections of this polygon and their tangents along the shape.
+     *
+     * This method returns all intersection tangents (as vectors) with this shape. The returned array of vectors is in no specific order.
+     *
+     * @param line
+     * @param inVectorBoundsOnly
+     * @returns
+     */
     Polygon.prototype.lineIntersectionTangents = function (line, inVectorBoundsOnly) {
         var _this = this;
         if (inVectorBoundsOnly === void 0) { inVectorBoundsOnly = false; }
@@ -477,6 +489,7 @@ var Polygon = /** @class */ (function () {
             return new Vector_1.Vector(polyLine.a.clone(), polyLine.b.clone()).moveTo(intersectionTuple.intersectionPoint);
         });
     };
+    //--- END --- Implement interface `Intersectable`
     /**
      * Get the closest line-polygon-intersection point (closest the line point A).
      *