+++ src/esm/interfaces/core.d.ts
@@ -36,6 +36,7 @@ import { PlotBoilerplate } from "../PlotBoilerplate";
 import { DrawLib } from "./DrawLib";
 import { PBText } from "../PBText";
 import { CSSBackdropFilterParams } from "./externals";
+import { VertTuple } from "../VertTuple";
 /**
  * @classdesc Coordinates (x,y) on the plane.
  *
@@ -216,4 +217,31 @@ export interface IDraggable {
     cindex: number;
     isVertex(): boolean;
     setVIndex(vindex: number): IDraggable;
+}
+/**
+ * Shapes should implement this interface if they support intersection calculation.
+ */
+export interface Intersectable {
+    /**
+     * Get all line intersections with this polygon.
+     *
+     * This method returns all intersections (as vertices) with this shape. The returned array of vertices is in no specific order.
+     *
+     * See demo `47-closest-vector-projection-on-polygon` for how it works.
+     *
+     * @param {VertTuple} line - The line to find intersections with.
+     * @param {boolean} inVectorBoundsOnly - (default=false) If set to true only intersecion points on the passed vector are returned (located strictly between start and end vertex).
+     * @returns {Array<Vertex>} - An array of all intersections within this shape's bounds.
+     */
+    lineIntersections(line: VertTuple<any>, inVectorBoundsOnly?: boolean): Array<Vertex>;
+    /**
+     * Get all line intersections of this polygon and their tangents along the shape.
+     *
+     * This method returns all intersection tangents (as vectors) with this shape. The returned array of vectors is in no specific order.
+     *
+     * @param {VertTuple} line - The line to find intersections with.
+     * @param  {boolean} inVectorBoundsOnly - (default=false) If set to true only intersecion points on the passed vector are returned (located strictly between start and end vertex).
+     * @returns {Array<Vector>} - An array of all intersection tangents within this shape's bounds.
+     */
+    lineIntersectionTangents(line: VertTuple<any>, inVectorBoundsOnly?: boolean): Array<Vector>;
 }